+++ e2e/cypress.config.js
@@ -3,7 +3,7 @@ import { defineConfig } from "cypress";
 export default defineConfig({
   viewportWidth: 1920,
   viewportHeight: 1080,
-  defaultCommandTimeout: 60 * 1000,
+  defaultCommandTimeout: 360 * 1000,
   // video: false,
   e2e: {
     setupNodeEvents(on, config) {